Task: Getting a photo into an app
This change provides basic functionality of getting a photo into an app.
It uses a Navigation Controller UI widget, and the SlViewController drives that interaction.
Currently:
- a default image will appear (spaceship.png) the first time that the app is loaded by the user.
- the user can use the 'camera' button on the top left to choose images from the photo gallery. Presumably, if a camera is present, the user can take a picture and load it in as well.
- once the picture is loaded, it is saved into the apps "Documents" folder. When the app is restarted and a saved image exists, it will be loaded (instead of the default 'spaceship.png').

This commit needs to be tested on another machine to ensure that all committed sources are also the required sources for compilation on another instance of Xcode.

Signed-off-by: Martin Lizon <martin_lizon@xyratex.com>